fix: expansion on quotes
