+++ srcs/input_expand.c
@@ -6,13 +6,13 @@
 /*   By: maalexan <maalexan@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/09 10:14:08 by maalexan          #+#    #+#             */
-/*   Updated: 2023/07/20 09:51:22 by maalexan         ###   ########.fr       */
+/*   Updated: 2023/07/21 22:58:09 by maalexan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-static void	copy_and_hide(t_env *var, char *dst, char *src)
+static void	copy_key_or_value(t_env *var, char *dst, char *src)
 {
 	if (!ft_strncmp(var->key, src, ft_strlen(src)))
 		*dst++ = '$';
@@ -44,7 +44,7 @@ static int	assess_len(char *str, int len, char **copy)
 		src = var_has_quote(variable);
 	str[len] = temp;
 	if (copy && src)
-		copy_and_hide(variable, *copy, src);
+		copy_key_or_value(variable, *copy, src);
 	if (src && !ft_strncmp(variable->key, src, ft_strlen(src)))
 		return (ft_strlen(src) + 1);
 	else if (src)
