+++ srcs/input_expand_utils.c
@@ -6,7 +6,7 @@
 /*   By: maalexan <maalexan@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/18 21:13:56 by maalexan          #+#    #+#             */
-/*   Updated: 2023/07/19 21:41:59 by maalexan         ###   ########.fr       */
+/*   Updated: 2023/07/22 00:21:17 by maalexan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,35 +56,42 @@ int	quote_closes(char *str)
 }
 
 /*
-**	Guarantees the string
-**	has only one quote
-**/
-int	has_single_quote(char *str)
-{
-	int	quote;
-
-	quote = 0;
-	while (*str)
-		if (is_quote(*str++))
-			quote++;
-	if (quote == 1)
-		return (1);
-	return (0);
-}
-
-/*
 **	Confirms that the string is an
 **	environment variable and that
-**	it has only one quote
+**	it's composition has quotes
 **/
 int	is_a_quoted_var(char *str)
 {
 	t_ctrl	*ctrl;
 	t_env	*var;
+	char	*cursor;
 
 	ctrl = get_control();
 	var = search_var(str, ctrl->env);
-	if (!var || !has_single_quote(var->value))
+	if (!var)
 		return (0);
-	return (1);
+	cursor = var->value;
+	while (*cursor)
+		if (is_quote(*cursor++))
+			return (1);
+	return (0);
+}
+
+/*
+**	Acessory function to help
+**	copy the whole quoted string
+*/
+
+int	goto_next_quote(char *args)
+{
+	int		i;
+	char	quote;
+
+	i = 1;
+	if (!quote_closes(args))
+		return (0);
+	quote = *args++;
+	while (*args++ != quote)
+		i++;
+	return (i);
 }
